backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow==2.20.0
keras==3.11.0